ffit/componente icon e text
atualizando o componente para trabalhar a obrigatoriedade das propriedades, conforme o tipo for escolhido.
